Add some boolean options.
onePerSide: true combines the left and bottom cutouts into one per side, false
is as before.

openBottom: true prints a bumper style case where the top and bottom cutouts
are symmetrical (as I initially designed it) and false leaves the entire bottom
covered except for the camera and adjacent bits.